Add dynamic robots.txt from API
