+++ .changeset/beige-feet-shake.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": minor
+---
+
+Add dynamic robots.txt from control panel settings
